Declare cast routines in cast.h
